refactor: now the book list always know about user
So all list contains the user "first letter" and renew works from any of
the lists.

OMG: the code is starting looking like a nightmare. I need to refactor
that, and maybe I need to switch away from reflex, perhaps the paradigm
is too complicated, or parhaps I'm too stupid ;)