feat: menu for actions, making them more understandable
And because I've not done another commit, I'm also cleaned the timing
logic. Now we have a `now` value which is updated in a dynamic every s
and which updates the refresh counter, the different days logic AND will
soon be able to automatically refresh every hour when the application is
opened.